[MC][TableGen] Store itinerary cycles as uint8_t
OperandCycles and ForwardingPaths contain values that fit in uint8_t. Store both generated tables as uint8_t, reject out-of-range OperandCycles and bypass counts, and emit size and range static_asserts.

In an AppleClang Release build, the five target table pairs shrink from 214,000 to 53,500 bytes. __TEXT,__const shrinks 160,496 bytes with unchanged __DATA_CONST,__const, __DATA,__data, bind counts, and rebase counts. Stripped llc and llvm-mca each shrink 148,592 bytes; the stripped llvm multicall binary shrinks 165,120 bytes.

A 16-pair Cortex-A9 llc benchmark changes CPU time by +0.272% with a 95% confidence interval of [-1.239%, +1.806%]. Baseline and candidate assembly is byte-identical. The full six-target build and all 420 TableGen tests pass.

+++ llvm/include/llvm/CodeGen/TargetSubtargetInfo.h
@@ -72,7 +72,7 @@ protected: // Can only create subclasses...
                       const MCWriteProcResEntry *WPR,
                       const MCWriteLatencyEntry *WL,
                       const MCReadAdvanceEntry *RA, const InstrStage *IS,
-                      const unsigned *OC, const unsigned *FP);
+                      const uint8_t *OC, const uint8_t *FP);
 
 public:
   // AntiDepBreakMode - Type of anti-dependence breaking that should
