+++ llvm/lib/CodeGen/TargetSubtargetInfo.cpp
@@ -19,7 +19,7 @@ TargetSubtargetInfo::TargetSubtargetInfo(
     ArrayRef<StringRef> PN, ArrayRef<SubtargetFeatureKV> PF,
     ArrayRef<SubtargetSubTypeKV> PD, const MCWriteProcResEntry *WPR,
     const MCWriteLatencyEntry *WL, const MCReadAdvanceEntry *RA,
-    const InstrStage *IS, const unsigned *OC, const unsigned *FP)
+    const InstrStage *IS, const uint8_t *OC, const uint8_t *FP)
     : MCSubtargetInfo(TT, CPU, TuneCPU, FS, PN, PF, PD, WPR, WL, RA, IS, OC,
                       FP) {}
